requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
pandas==2.3.3
pyotp==2.9.0
python-dateutil==2.9.0.post0
protobuf==6.33.2
requests==2.32.5
six==1.17.0
smartapi-python==1.5.5
urllib3==2.6.3
websocket-client==1.9.0
